# dist/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação da API Task Manager",
        "version": "1.0.0",
        "description": "Documentação da API utilizando Swagger"
    },
    "paths": {
        "/users": {
            "post": {
                "summary": "Criar um novo usuário",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "required": ["name", "email", "password"]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso"
                    },
                    "500": {
                        "description": "Erro ao criar usuário"
                    }
                }
            },
            "get": {
                "summary": "Listar todos os usuários",
                "responses": {
                    "200": {
                        "description": "Lista de usuários obtida com sucesso"
                    },
                    "500": {
                        "description": "Erro ao obter lista de usuários"
                    }
                }
            }
        },
        "/users/{userId}": {
            "delete": {
                "summary": "Excluir usuário por ID",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID do usuário a ser excluído",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário excluído com sucesso"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    },
                    "500": {
                        "description": "Erro ao excluir usuário"
                    }
                }
            }
        },
        "/users/{projectId}/add-members/{userId}": {
            "post": {
                "summary": "Adicionar usuário inicial a um projeto",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "ID do projeto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID do usuário a ser adicionado ao projeto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário adicionado com sucesso ao projeto"
                    },
                    "400": {
                        "description": "Erro ao adicionar usuário ao projeto"
                    }
                }
            }
        },
        "/projects": {
            "post": {
                "summary": "Criar um novo projeto",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "creatorId": {
                                        "type": "integer"
                                    }
                                },
                                "required": ["name", "description", "creatorId"]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Projeto criado com sucesso"
                    },
                    "500": {
                        "description": "Erro ao criar projeto"
                    }
                }
            },
            "get": {
                "summary": "Listar todos os projetos",
                "responses": {
                    "200": {
                        "description": "Lista de projetos obtida com sucesso"
                    },
                    "500": {
                        "description": "Erro ao obter lista de projetos"
                    }
                }
            }
        },
        "/projects/{projectId}": {
            "delete": {
                "summary": "Excluir projeto por ID",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "ID do projeto a ser excluído",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Projeto excluído com sucesso"
                    },
                    "404": {
                        "description": "Projeto não encontrado"
                    },
                    "500": {
                        "description": "Erro ao excluir projeto"
                    }
                }
            }
        },
        "/projects/{projectId}/add-members/{userId}": {
            "post": {
                "summary": "Adicionar usuário como membro a um projeto",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "ID do projeto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID do usuário a ser adicionado como membro do projeto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "creatorId": {
                                        "type": "integer"
                                    }
                                },
                                "required": ["creatorId"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário adicionado ao projeto com sucesso"
                    },
                    "400": {
                        "description": "Erro ao adicionar usuário ao projeto"
                    }
                }
            }
        },
        "/projects/{projectId}/remove-members/{userId}": {
            "delete": {
                "summary": "Remover usuário de um projeto",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "ID do projeto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID do usuário a ser removido do projeto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário removido do projeto com sucesso"
                    },
                    "500": {
                        "description": "Erro ao remover usuário do projeto"
                    }
                }
            }
        },
        "/tasks": {
            "post": {
                "summary": "Criar uma nova tarefa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "userId": {
                                        "type": "integer"
                                    },
                                    "projectId": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "required": ["userId", "projectId", "title"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarefa criada com sucesso"
                    },
                    "500": {
                        "description": "Erro ao criar tarefa"
                    }
                }
            }
        },
        "/tasks/{id}": {
            "put": {
                "summary": "Atualizar uma tarefa pelo ID",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "ID da tarefa a ser atualizada",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "userId": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "required": ["userId", "title", "status"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarefa atualizada com sucesso"
                    },
                    "500": {
                        "description": "Erro ao atualizar tarefa"
                    }
                }
            }
        },
        "/tasks/{id}/status": {
            "put": {
                "summary": "Atualizar o status de uma tarefa pelo ID",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "ID da tarefa a ter o status atualizado",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "newStatus": {
                                        "type": "string"
                                    }
                                },
                                "required": ["newStatus"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status da tarefa atualizado com sucesso"
                    },
                    "500": {
                        "description": "Erro ao atualizar o status da tarefa"
                    }
                }
            }
        },
        "/tags": {
            "get": {
                "summary": "Listar todas as tags",
                "responses": {
                    "200": {
                        "description": "Lista de tags obtida com sucesso"
                    },
                    "500": {
                        "description": "Erro ao buscar as tags"
                    }
                }
            },
            "post": {
                "summary": "Criar uma nova tag",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    }
                                },
                                "required": ["title"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tag criada com sucesso"
                    },
                    "500": {
                        "description": "Erro ao criar a tag"
                    }
                }
            }
        },
        "/tags/{id}": {
            "get": {
                "summary": "Obter uma tag pelo ID",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "ID da tag a ser obtida",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tag obtida com sucesso"
                    },
                    "404": {
                        "description": "Tag não encontrada"
                    },
                    "500": {
                        "description": "Erro ao buscar a tag"
                    }
                }
            },
            "put": {
                "summary": "Atualizar uma tag pelo ID",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "ID da tag a ser atualizada",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    }
                                },
                                "required": ["title"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tag atualizada com sucesso"
                    },
                    "404": {
                        "description": "Tag não encontrada"
                    },
                    "500": {
                        "description": "Erro ao atualizar a tag"
                    }
                }
            },
            "delete": {
                "summary": "Excluir uma tag pelo ID",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "ID da tag a ser excluída",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tag excluída com sucesso"
                    },
                    "500": {
                        "description": "Erro ao excluir a tag"
                    }
                }
            }
        }
    }
}